\name{GSE20986_design}
\docType{data}
\alias{GSE20986_design}
\title{Comparative Gene Expression Profiling of HUVEC and Ocular Vascular Endothelial Cells (GEO database entry GSE20986) microarray design data file}
\description{
	Evolution combined with genomic study elucidates genetic
	bases of isobutanol tolerance in Escherichia coli

	\bold{Platform:}          GPL570 [HG-U133_Plus_2] Affymetrix Human Genome U133 Plus 2.0 Array \cr
	\bold{Organism:}          Homo sapiens \cr
	\bold{Experiment type:}   Expression profiling by array \cr
	\bold{Summary} \cr

	To compare the gene expression profiles of unpassaged, proliferating HUVEC and
	human iris, retinal and choroidal microvascular endothelial cells.
	Gene expression profiling revealed significant differences between HUVEC and
	ocular microvascular endothelial cells suggesting that HUVE cells may not be a
	suitable surrogate when studying pathophysiological mechanisms of ocular
	disorders. There were significant differences in the gene expression of important
	cell signalling pathways in human retinal and choroidal ECs. These differences may
	be important in the mechanisms and treatment of choroidal and retinal
	neovascularisation.

	\bold{Overall design} \cr
	12 arrays are included. Endothelial cells were derived from 4 tissues: iris,
	retina, choroid and human umbilical vein. RNA extracts from cells were hybridised
	to Affymetrix HGU133plus2 arrays in triplicate.

	\bold{Samples} \cr
	\itemize{
		\item{GSM524662 --- iris-cellline-1}
		\item{GSM524663 --- retina-cellline-1}
		\item{GSM524664 --- retina-cellline-2}
		\item{GSM524665 --- iris-cellline-2}
		\item{GSM524666 --- retina-cellline-3}
		\item{GSM524667 --- iris-cellline-3}
		\item{GSM524668 --- choroid-cellline-1}
		\item{GSM524669 --- choroid-cellline-2}
		\item{GSM524670 --- choroid-cellline-3}
		\item{GSM524671 --- huvec-cellline-1}
		\item{GSM524672 --- huvec-cellline-2}
		\item{GSM524673 --- huvec-cellline-3}
	}
}
\usage{GSE20986_design}
\format{
	A 4 column table relating each sample (see above list) to parts of the body.
	The columns (parts of the body) are the following:
	\itemize{
		\item{choroid}
		\item{huvec}
		\item{iris}
		\item{retina}
	}
}
\source{Geo Accession #Series GSE20986}

\references{Browning, A. C., Halligan, E. P., Swan, D. C., et al. (Jan 2012) \dQuote{Comparative Gene Expression Profiling of HUVEC and Ocular Vascular Endothelial Cells.} \emph{Br J Ophthalmol.} Vol. \bold{96}, No. 1, p. 128--132. doi: 10.1136/bjophthalmol-2011-300572.}
\keyword{datasets}
