% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\docType{data}
\name{profiles}
\alias{profiles}
\title{Profiles Dataset}
\format{
A data frame with 31 rows and 5 columns:
\describe{
\item{x}{A sequence from 0 to 30 (x-Axis)}
\item{concave_down_pos_slo}{First concave profile (y-Axis): Concave down (positive slope)}
\item{concave_down_neg_slo}{Second concave profile (y-Axis): Concave down (negative slope)}
\item{concave_up_pos_slo}{Third concave profile (y-Axis): Concave up (positive slope)}
\item{concave_up_neg_slo}{Fourth concave profile (y-Axis): Concave up (negative slope)}
}
}
\usage{
profiles
}
\description{
A data frame describing four concave profiles along an x-Axis.
}
\keyword{datasets}
